翻新投稿提示，增加控制删除取消投稿后投稿消息能力

+++ src/bot/callback/users.py
@@ -2,6 +2,7 @@ from asyncio import sleep
 
 from telegram import Update
 from telegram.ext import ContextTypes, ConversationHandler
+from src.config import Config_submit
 
 
 async def cancel(update: Update, context: ContextTypes.DEFAULT_TYPE):
@@ -9,13 +10,13 @@ async def cancel(update: Update, context: ContextTypes.DEFAULT_TYPE):
     if query:
         await query.edit_message_text(text="操作已取消")
         origin_message = update.effective_message.reply_to_message
-        if origin_message:
+        if origin_message and Config_submit.SUBMIT_DELETE_WHEN_CANCEL:
             await context.bot.delete_message(
                 chat_id=origin_message.chat.id,
                 message_id=origin_message.message_id
             )
         await query.answer("操作已取消")
-    await sleep(1)
+    await sleep(2)
     await context.bot.delete_message(
         chat_id=update.effective_chat.id,
         message_id=update.effective_message.message_id
